Refactor/demoonlighting

+++ src/main/resources/liquibase/changes/more_partner_customization.sql
@@ -2,4 +2,4 @@ alter table route
     add column if not exists partner_code varchar;
 
 alter table route
-    add column if not exists estimated_duration_hrs double precision;
+    add column if not exists estimated_duration_hrs float;
